Cambios en los inputs
